# src/main/resources/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Clinica Médica API
  version: 1.0.0
  description: API da clínica médica com autenticação JWT.
servers:
  - url: http://localhost:8080
    description: Servidor local
paths:
  /api/auth/login:
    post:
      summary: Autentica um usuário e retorna o token JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciais inválidas.
  /api/auth/register:
    post:
      summary: Registra um novo usuário e retorna o token JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuário registrado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Dados inválidos.
        '409':
          description: Já existe usuário com as mesmas credenciais.
...
components:
...
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - senha
      properties:
        email:
          type: string
          format: email
        senha:
          type: string
    RegisterRequest:
      type: object
      required:
        - nome
        - cpf
        - idade
        - email
        - senha
      properties:
        nome:
          type: string
        cpf:
          type: string
          description: 11 dígitos numéricos
        idade:
          type: integer
          format: int32
          minimum: 0
        email:
          type: string
          format: email
        senha:
          type: string
          format: password
        roles:
          type: array
          description: Perfis de acesso a serem associados ao usuário
          items:
            type: string
            enum:
              - ADMIN
              - MEDICO
              - RECEPCIONISTA
    AuthResponse:
      type: object
      required:
        - token
        - expiresAt
        - tokenType
        - user
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        tokenType:
          type: string
        user:
          $ref: '#/components/schemas/UsuarioResumo'
    UsuarioResumo:
      type: object
      required:
        - id
        - nome
        - email
        - roles
      properties:
        id:
          type: string
        nome:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
            enum:
              - ADMIN
              - MEDICO
              - RECEPCIONISTA
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT